src/requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
django_rest_framework
parse>=1.19.0
pyyaml
django-bootstrap-v5
stripe
python-dotenv
